Rename of OAS examples
