+++ aviary/examples/external_subsystems/OAS_mass/test/test_OAS_wing_weight_analysis.py
@@ -4,14 +4,14 @@ import numpy as np
 import openmdao.api as om
 from openmdao.utils.testing_utils import use_tempdirs
 
-from aviary.examples.external_subsystems.OAS_weight.OAS_wing_weight_analysis import OAStructures
+from aviary.examples.external_subsystems.OAS_mass.OAS_wing_mass_analysis import OAStructures
 
 
 class Test_OAStructures(unittest.TestCase):
-    """Test OAS wing weight component."""
+    """Test OAS wing mass component."""
 
     @use_tempdirs
-    def test_OAS_wing_weight_analysis(self):
+    def test_OAS_wing_mass_analysis(self):
         # run program
         prob = om.Problem()
